serde: move version typedef to separate header
