+++ src/v/serde/version.h
@@ -1,0 +1,28 @@
+// Copyright 2023 Redpanda Data, Inc.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.md
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0
+
+#pragma once
+
+#include <cinttypes>
+
+namespace serde {
+
+using version_t = std::uint8_t;
+
+template<version_t V>
+struct version {
+    static constexpr auto const v = V;
+};
+
+template<version_t V>
+struct compat_version {
+    static constexpr auto const v = V;
+};
+
+} // namespace serde
